Make sure that XMLPARSEAPI specifies the calling convention when building under MSVC -- this is needed when using the pre-compiled DLL with projects built using a different calling convention.
XMLPARSEAPI now takes the return type as a parameter and inserts annotations
on both sides of the type to make sure the compiler is happy.  A new
macro, XMLCALLBACK, is used to perform similar annotation of the callback
function types, which do not need the dllimport/dllexport annotations but
do still need the __cdecl annotation.

This closes SF bug #413653.